Split character builder server helpers
