+++ frontend/src/server/tools/character-builder/prompt.ts
@@ -1,0 +1,288 @@
+import {
+  getAccessoryPrompt,
+  getDistinctiveFeaturePrompt,
+  getRealismPrompt,
+  getTraitOptionPrompt,
+} from '@/lib/character-builder';
+import type {
+  CharacterBuilderConsistencyMode,
+  CharacterBuilderReferenceImage,
+  CharacterBuilderReferenceStrength,
+  CharacterBuilderRequest,
+  CharacterBuilderTraits,
+} from '@/types/character-builder';
+import {
+  buildAnchorParts,
+  buildCloseUpIdentityParts,
+  buildLightingPreservationParts,
+  buildPreservationParts,
+  getCustomHairDescription,
+  getCustomOutfitDescription,
+  isHairEnabled,
+  isOutfitEnabled,
+} from './traits';
+import { collectReferenceImages, trimString, uniqueStrings } from './utils';
+
+function buildReferenceStrengthBlock(
+  referenceStrength: CharacterBuilderReferenceStrength | null,
+  traits: CharacterBuilderTraits
+): string {
+  const anchorParts = buildAnchorParts(traits);
+  switch (referenceStrength) {
+    case 'loose':
+      return 'Use the reference image for overall vibe and recognizable identity cues, with room to clean up the design.';
+    case 'strong':
+      return `Stay very close to the uploaded identity reference for ${anchorParts} and defining visible markers.`;
+    case 'balanced':
+    default:
+      return 'Stay close to the uploaded identity reference while cleaning framing, lighting, and readability.';
+  }
+}
+
+function buildConsistencyBlock(
+  consistencyMode: CharacterBuilderConsistencyMode,
+  traits: CharacterBuilderTraits
+): string {
+  switch (consistencyMode) {
+    case 'exploratory':
+      return 'Keep the same person recognizable while allowing a little stylistic variation.';
+    case 'strict':
+      return `Preserve the same identity, ${buildAnchorParts(traits)}, and distinctive markers as consistently as possible.`;
+    case 'balanced':
+    default:
+      return 'Aim for stable identity and styling while still allowing small cleanup improvements.';
+  }
+}
+
+function buildTraitPrompt(traits: CharacterBuilderTraits, sourceMode: CharacterBuilderRequest['sourceMode']): string[] {
+  const parts: string[] = [];
+  const pushValue = (value: string | null) => {
+    if (value) parts.push(value);
+  };
+  const pushTrait = (
+    key: keyof Pick<
+      CharacterBuilderTraits,
+      | 'genderPresentation'
+      | 'ageRange'
+      | 'skinTone'
+      | 'faceCues'
+      | 'hairColor'
+      | 'hairLength'
+      | 'hairstyle'
+      | 'eyeColor'
+      | 'bodyBuild'
+      | 'outfitStyle'
+    >
+  ) => {
+    const trait = traits[key];
+    if (!trait || typeof trait !== 'object' || !('value' in trait)) return;
+    if (trait.value == null || trait.value === 'auto') return;
+    if (key === 'genderPresentation' && trait.value === 'custom') {
+      const customGenderDescription = trimString(traits.customGenderDescription);
+      if (customGenderDescription) {
+        parts.push(customGenderDescription);
+      }
+      return;
+    }
+    pushValue(getTraitOptionPrompt(key, trait.value));
+  };
+
+  pushTrait('genderPresentation');
+  pushTrait('ageRange');
+  pushTrait('skinTone');
+  pushTrait('faceCues');
+  if (isHairEnabled(traits)) {
+    const customHairDescription = getCustomHairDescription(traits);
+    if (customHairDescription) {
+      parts.push(customHairDescription);
+    } else {
+      pushTrait('hairColor');
+      pushTrait('hairLength');
+      pushTrait('hairstyle');
+    }
+  }
+  pushTrait('eyeColor');
+  pushTrait('bodyBuild');
+  if (isOutfitEnabled(traits)) {
+    const customOutfitDescription = getCustomOutfitDescription(traits);
+    if (customOutfitDescription) {
+      parts.push(customOutfitDescription);
+    } else {
+      pushTrait('outfitStyle');
+    }
+  }
+
+  traits.accessories.forEach((value) => pushValue(getAccessoryPrompt(value)));
+  traits.distinctiveFeatures.forEach((value) => pushValue(getDistinctiveFeaturePrompt(value)));
+  if (trimString(traits.customDetailsDescription)) {
+    pushValue(trimString(traits.customDetailsDescription));
+  }
+  pushValue(getRealismPrompt(traits.realismStyle));
+
+  if (sourceMode === 'reference-image') {
+    return parts;
+  }
+
+  return parts;
+}
+
+function buildReferenceBlock(input: CharacterBuilderRequest, referenceImages: CharacterBuilderReferenceImage[]): string[] {
+  const identityImages = collectReferenceImages(referenceImages, 'identity');
+  const styleImages = collectReferenceImages(referenceImages, 'style');
+  const blocks: string[] = [];
+
+  if (identityImages.length) {
+    blocks.push(
+      `Treat the uploaded identity reference image as the primary anchor for ${buildAnchorParts(input.traits)} and recognizable identity cues.`
+    );
+    blocks.push(buildReferenceStrengthBlock(input.referenceStrength ?? 'balanced', input.traits));
+  }
+
+  if (styleImages.length) {
+    blocks.push(
+      isOutfitEnabled(input.traits)
+        ? 'Use any secondary reference image only for outfit, palette, or styling inspiration. Do not replace the core identity with it.'
+        : 'Use any secondary reference image only for palette or overall art direction. Do not replace the core identity with it.'
+    );
+  }
+
+  if (identityImages.length && input.sourceMode === 'reference-image') {
+    blocks.push('Apply manual builder choices as supporting constraints rather than hard replacements.');
+  }
+
+  return blocks;
+}
+
+function buildLayoutBlock(input: CharacterBuilderRequest): string[] {
+  const { action, outputMode, outputOptions } = input;
+
+  if (action === 'full-body-fix') {
+    return [
+      'Rebuild the character as a clean head-to-toe reference image.',
+      'Ensure the full head, torso, arms, hands, legs, ankles, and feet are visible with no crop.',
+      `Preserve the same ${buildPreservationParts(input.traits)} from the source image.`,
+    ];
+  }
+
+  if (action === 'lighting-variant') {
+    return [
+      'Create a lighting-only variation of the selected character reference.',
+      `Preserve the same ${buildLightingPreservationParts(input.traits)}.`,
+      'Change the lighting mood while keeping the character readable and reference-friendly.',
+    ];
+  }
+
+  if (outputMode === 'character-sheet') {
+    const blocks = [
+      'Create a clean multi-angle character sheet with front, three-quarter, side, and back views.',
+      'Keep scale, lighting, anatomy, and spacing consistent across the sheet.',
+    ];
+    if (outputOptions.includeCloseUps) {
+      blocks.push('Arrange the sheet as a clean two-row grid.');
+      blocks.push('Top row: four full-body views showing front, side, three-quarter, and back.');
+      blocks.push('Reserve roughly two-thirds of the total canvas height for the top row so each full-body panel is tall enough to show the complete head-to-toe figure.');
+      blocks.push('Each top-row panel must include the full head, full legs, ankles, and feet with no crop.');
+      blocks.push('Bottom row: four compact close-up head-and-shoulders portraits directly underneath, matching the same view order.');
+      blocks.push('Keep the bottom row to roughly one-third of the total canvas height so the close-ups read as a shorter strip and do not steal height from the full-body views.');
+      blocks.push('Keep all eight panels aligned with even spacing, the same neutral background, and consistent lighting.');
+      blocks.push(
+        `The close-ups must clearly show the ${buildCloseUpIdentityParts(input.traits)} as the full-body views.`
+      );
+    }
+    return blocks;
+  }
+
+  const blocks = [
+    'Create a clean front-facing character reference portrait framed as a balanced medium portrait, with strong facial clarity and readable hair and signature details.',
+    'The face must be clearly visible and facing the camera, but avoid an extreme close-up that crops too tightly around the face.',
+    'Frame from roughly mid-torso or chest-up, showing the full head, neck, shoulders, and upper torso. Do not produce a full-body pose, a standing wide shot, or a fashion/editorial composition.',
+    'Bias toward a stable, reusable identity anchor rather than a generic beauty shot.',
+  ];
+  if (outputMode !== 'portrait-reference' && outputOptions.fullBodyRequired) {
+    blocks.push('Frame the character full body from head to toe.');
+  }
+  return blocks;
+}
+
+function buildOutputOptionBlock(input: CharacterBuilderRequest): string[] {
+  const blocks: string[] = [];
+  if (input.outputOptions.neutralStudioBackground) {
+    blocks.push('Use a clean neutral studio background.');
+  }
+  if (input.outputOptions.preserveFacialDetails) {
+    blocks.push('Preserve facial detail clarity and readable skin texture.');
+  }
+  if (input.outputOptions.avoid3dRenderLook) {
+    blocks.push('Avoid a glossy 3D-render look.');
+  }
+  if (input.outputMode === 'character-sheet') {
+    blocks.push('Keep the character fully visible with realistic anatomy and complete framing.');
+  }
+  return blocks;
+}
+
+function buildDetailLockBlock(input: CharacterBuilderRequest): string[] {
+  const detailTags = uniqueStrings(input.mustRemainVisible ?? []);
+  const blocks: string[] = [];
+
+  if (input.traits.distinctiveFeatures.length) {
+    blocks.push(
+      `Preserve these distinctive traits: ${input.traits.distinctiveFeatures
+        .map((value) => getDistinctiveFeaturePrompt(value) ?? value)
+        .filter(Boolean)
+        .join(', ')}.`
+    );
+  }
+
+  if (trimString(input.traits.customDetailsDescription)) {
+    blocks.push(`Preserve these extra visible details: ${trimString(input.traits.customDetailsDescription)}.`);
+  }
+
+  if (detailTags.length) {
+    blocks.push(`These details must remain visible: ${detailTags.join(', ')}.`);
+  }
+
+  return blocks;
+}
+
+export function buildPrompt(input: CharacterBuilderRequest): string {
+  const referenceImages = input.referenceImages.filter((image) => /^https?:\/\//i.test(image.url));
+  const sections: string[] = [
+    'Create a reusable AI character reference asset as a single still image output. Do not imply motion, timeline editing, or video generation.',
+  ];
+  const traitPrompt = buildTraitPrompt(input.traits, input.sourceMode);
+  const referenceBlock = buildReferenceBlock(input, referenceImages);
+  const layoutBlock = buildLayoutBlock(input);
+  const outputBlock = buildOutputOptionBlock(input);
+  const detailLockBlock = buildDetailLockBlock(input);
+  const advancedNotes = trimString(input.advancedNotes);
+
+  if (traitPrompt.length) {
+    sections.push(`Character DNA: ${traitPrompt.join(', ')}.`);
+  } else if (input.sourceMode === 'scratch') {
+    sections.push('Create an original character identity using the selected builder traits and the output constraints below.');
+  } else {
+    sections.push('Infer the core identity from the uploaded reference image while following the selected builder constraints.');
+  }
+
+  if (referenceBlock.length) {
+    sections.push(referenceBlock.join(' '));
+  }
+
+  sections.push(buildConsistencyBlock(input.consistencyMode, input.traits));
+  sections.push(layoutBlock.join(' '));
+
+  if (outputBlock.length) {
+    sections.push(outputBlock.join(' '));
+  }
+
+  if (detailLockBlock.length) {
+    sections.push(detailLockBlock.join(' '));
+  }
+
+  if (advancedNotes) {
+    sections.push(`Additional guidance: ${advancedNotes}.`);
+  }
+
+  return sections.join('\n\n');
+}
