+++ frontend/src/server/tools/character-builder/types.ts
@@ -1,0 +1,5 @@
+import type { CharacterBuilderRequest } from '@/types/character-builder';
+
+export type RunCharacterBuilderInput = CharacterBuilderRequest & {
+  userId: string;
+};
